% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRRR.R
\name{SRRR2}
\alias{SRRR2}
\title{SRRR2}
\usage{
SRRR2(X, Y, lambda, r)
}
\arguments{
\item{X}{X is multiple predictor matrix}

\item{Y}{Y is multiple response matrix}

\item{lambda}{lambda is penalty parameters vector}

\item{r}{r is a constraint on the rank}
}
\value{
a list(two matrix A,B and a index)
}
\description{
SRRR function,use variational method .
}
\details{
This is some descriptio of this function.

you can use this function to caculate SRRR estimator.
}
