\name{StatComp21058-package}
\alias{StatComp21058-package}
\alias{StatComp21058}
\docType{package}
\title{
  StatComp21058
}
\description{
  all my homework and two SRRR functions.
}
\details{
  the two function use subgradinets method and variational method.
}
\author{
guo yuanheng

Maintainer: guo yuanheng <429555967@qq.com>
}
\references{
  Sparse Reduced-Rank Regression for Simultaneous
Dimension Reduction and Variable Selection
Lisha Chen & Jianhua Z. Huang
}
\keyword{ package }